Finish reintroducing DID resolvers
