+++ packages/dids/tests/resolver/did-resolver.spec.ts
@@ -1,13 +1,13 @@
-import * as sinon from 'sinon';
-import { expect } from 'chai';
+import type { UnwrapPromise } from '@web5/common';
 
-// import type { DidResolverCache } from '../../src/resolver/did-resolver.js';
+import { expect } from 'chai';
+import * as sinon from 'sinon';
 
 import { DidJwk } from '../../src/methods/did-jwk.js';
 import { DidResource } from '../../src/types/did-core.js';
 import { isDidVerificationMethod } from '../../src/utils.js';
 import { DidResolver } from '../../src/resolver/did-resolver.js';
-// import { didResolverTestVectors } from './fixtures/test-vectors/did-resolver.js';
+import DidJwkResolveTestVector from '../../../../test-vectors/did_jwk/resolve.json' assert { type: 'json' };
 
 describe('DidResolver', () => {
   describe('resolve()', () => {
@@ -17,13 +17,6 @@ describe('DidResolver', () => {
       const didMethodApis = [DidJwk];
       didResolver = new DidResolver({ didResolvers: didMethodApis });
     });
-
-    // it('passes test vectors', async () => {
-    //   for (const vector of didResolverTestVectors) {
-    //     const didResolutionResult = await didResolver.resolve(vector.input);
-    //     expect(didResolutionResult.didDocument).to.deep.equal(vector.output);
-    //   }
-    // });
 
     it('returns an invalidDid error if the DID cannot be parsed', async () => {
       const didResolutionResult = await didResolver.resolve('unparseable:did');
@@ -45,74 +38,20 @@ describe('DidResolver', () => {
       expect(didResolutionResult.didResolutionMetadata).to.have.property('error', 'methodNotSupported');
     });
 
-    // describe('with LevelDB cache', () => {
-    //   let cache: DidResolverCache;
+    it('pass DID JWK resolve test vectors', async () => {
+        type TestVector = {
+          description: string;
+          input: Parameters<typeof DidJwk.resolve>[0];
+          output: UnwrapPromise<ReturnType<typeof DidJwk.resolve>>;
+          errors: boolean;
+        };
 
-    //   before(() => {
-    //     cache = new DidResolverCacheLevel();
-    //   });
+        for (const vector of DidJwkResolveTestVector.vectors as unknown as TestVector[]) {
+          const didResolutionResult = await DidJwk.resolve(vector.input);
 
-    //   beforeEach(async () => {
-    //     await cache.clear();
-    //     const didMethodApis = [DidKeyMethod];
-    //     didResolver = new DidResolver({ cache, didResolvers: didMethodApis });
-    //   });
-
-    //   after(async () => {
-    //     await cache.clear();
-    //   });
-
-    //   it('should cache miss for the first resolution attempt', async () => {
-    //     const did = 'did:key:z6MkmNvXGmVuux5W63nXKEM8zoxFmDLNfe7siCKG2GM7Kd8D';
-    //     // Create a Sinon spy on the get method of the cache
-    //     const cacheGetSpy = sinon.spy(cache, 'get');
-
-    //     await didResolver.resolve(did);
-
-    //     // Verify that cache.get() was called.
-    //     expect(cacheGetSpy.called).to.be.true;
-
-    //     // Verify the cache returned undefined.
-    //     const getCacheResult = await cacheGetSpy.returnValues[0];
-    //     expect(getCacheResult).to.be.undefined;
-
-    //     cacheGetSpy.restore();
-    //   });
-
-    //   it('should cache hit for the second resolution attempt', async () => {
-    //     const did = 'did:key:z6MkmNvXGmVuux5W63nXKEM8zoxFmDLNfe7siCKG2GM7Kd8D';
-    //     // Create a Sinon spy on the get method of the cache
-    //     const cacheGetSpy = sinon.spy(cache, 'get');
-    //     const cacheSetSpy = sinon.spy(cache, 'set');
-
-    //     await didResolver.resolve(did);
-
-    //     // Verify there was a cache miss.
-    //     expect(cacheGetSpy.calledOnce).to.be.true;
-    //     expect(cacheSetSpy.calledOnce).to.be.true;
-
-    //     // Verify the cache returned undefined.
-    //     let getCacheResult = await cacheGetSpy.returnValues[0];
-    //     expect(getCacheResult).to.be.undefined;
-
-    //     // Resolve the same DID again.
-    //     await didResolver.resolve(did);
-
-    //     // Verify that cache.get() was called.
-    //     expect(cacheGetSpy.called).to.be.true;
-    //     expect(cacheGetSpy.calledTwice).to.be.true;
-
-    //     // Verify there was a cache hit this time.
-    //     getCacheResult = await cacheGetSpy.returnValues[1];
-    //     expect(getCacheResult).to.not.be.undefined;
-    //     expect(getCacheResult).to.have.property('@context');
-    //     expect(getCacheResult).to.have.property('didDocument');
-    //     expect(getCacheResult).to.have.property('didDocumentMetadata');
-    //     expect(getCacheResult).to.have.property('didResolutionMetadata');
-
-    //     cacheGetSpy.restore();
-    //   });
-    // });
+          expect(didResolutionResult).to.deep.equal(vector.output);
+        }
+    });
   });
 
   describe('dereference()', () => {
@@ -123,11 +62,18 @@ describe('DidResolver', () => {
       didResolver = new DidResolver({ didResolvers: didMethodApis });
     });
 
-    it('returns a result with contentStream set to null and dereferenceMetadata.error set if resolution fails', async () => {
+    it('returns a result with contentStream set to null and dereferenceMetadata.error set to invalidDidUrl, if the DID URL is invalid', async () => {
       const result = await didResolver.dereference('abcd123;;;');
       expect(result.contentStream).to.be.null;
       expect(result.dereferencingMetadata.error).to.exist;
       expect(result.dereferencingMetadata.error).to.equal('invalidDidUrl');
+    });
+
+    it('returns a result with contentStream set to null and dereferenceMetadata.error set to invalidDid, if the DID is invalid', async () => {
+      const result = await didResolver.dereference('did:jwk:abcd123');
+      expect(result.contentStream).to.be.null;
+      expect(result.dereferencingMetadata.error).to.exist;
+      expect(result.dereferencingMetadata.error).to.equal('invalidDid');
     });
 
     it('returns a DID verification method resource as the value of contentStream if found', async () => {
